// src/types/pages.ts
export type Pages =
  | "demands"
  | "institutions"
  | "users"
  | "calendar"
  | "locales"
  | "releases"
  | "checklists"
  | "wiki"
  | "issues"
  | "chat"
  | "dashboard"
  | "documents"
  | "practices"
  | "objects"
  | "skills"
  | "labs"
  | "profile"
  | "login"
  | "create"
  | "show"
  | "archived"
  | "edit";